+++ kind-rational/lib/KindRational.hs
@@ -17,6 +17,7 @@
 module KindRational {--}
   ( -- * Rational kind
     Rational
+  , rational
   , type (%)
   , type (/)
   , Num
@@ -24,6 +25,7 @@ module KindRational {--}
   , showsPrecTypeLit
   , readPrecTypeLit
   , Normalize
+  , normalize
 
     -- * Types ⇔ Terms
   , KnownRational
@@ -116,7 +118,7 @@ pattern UnsafeNormalizedRational n d <-
 
 pattern_UnsafeNormalizedRational :: HasCallStack => P.Rational -> P.Rational
 pattern_UnsafeNormalizedRational = \(n P.:% d) ->
-  case mkRational n d of
+  case rational n d of
     Left e  -> error ("KindRational: " <> show e)
     Right r -> r
 
@@ -141,22 +143,27 @@ readPrecTypeLit = Read.parens $ do
     n :: P.Integer <- Read.parens $ ReadPrec.step I.readPrecTypeLit
     Read.expectP (Read.Symbol "%")
     d :: Natural <- Read.parens $ ReadPrec.step $ ReadPrec.lift pNatural
-    either fail pure $ mkRational n (toInteger d)
+    either fail pure $ rational n (toInteger d)
   where
     pNatural :: ReadP.ReadP Natural
     pNatural = read <$> ReadP.munch1 (\c -> c >= '0' && c <= '9')
 
--- | Creates a 'Prelude'.'P.Rational' using the given literal numerator
--- and denominator, which are expected to be already normalized.
-mkRational :: P.Integer -> P.Integer -> Either String P.Rational
-mkRational = \n d -> do
+-- | Creates a "Prelude".'P.Rational' using the given literal numerator
+-- and denominator, if they are already normalized.
+rational :: P.Integer -> P.Integer -> Either String P.Rational
+rational = \n d -> do
     when (d == 0) $ Left "Denominator is zero"
     when (d < 0) $ Left notNormMsg
     let r@(n' P.:% d') = n P.% d
     when (n /= n' || d /= d') $ Left notNormMsg
     Right r
   where
-    notNormMsg = "Rational is not normalized. Use Data.Ratio.% to normalize it."
+    notNormMsg = "Rational is not normalized. Use Data.Ratio.% or \
+                 \KindRational.normalize to normalize it."
+
+-- | Normalizes a "Prelude".'P.Rational'. 'Nothing' if denominator is zero.
+normalize :: P.Rational -> Maybe P.Rational
+normalize = \(n P.:% d) -> (n P.% d) <$ guard (d /= 0)
 
 --------------------------------------------------------------------------------